TaskSketcher_xxx.ui: use sensible minimal heights to avoid scaling problems
The problem is currently that if the user collapses the solver messages, the list views get too small. The reason is that there is no proper minimal height set. See the forum thread for screenshots of the fixes.

TaskSketcherMessages.ui: remove unnecessary space (saves also vertical space on small screens)