FEM: Restore path functions of run module
The functions getTempDir, getBesideDir and getBesideBase where moved to
femutils, modified and made public in the past. This breaks the solver
framework under certain circumstances. These folders are only intended
for the solver. Using it for other puroses may result in unexpected
behaviour (including crashes). This commit restores the original
functions for the solver module. The problem that other modules of the
workbench use those folders should be resolved at some point.