Path: allow boundaryDressup negative Exts
for effective use of boundaryDressup it needs to be able to set boundary limits which are not restrained to be >= work piece stock. Reuse of the Stock code means it gets the ( unnecessary ) limitations that stock must be greater than a rect/cyl around the 3D model. 

This mod frees both by allowing negative values to be input for Ext*neg and Ext*pos values

https://forum.freecadweb.org/viewtopic.php?f=15&t=41555