[Tools] remove ThumbnailProvider.sln
This single files does not help anything, one would also need the *.vcxproj files. But whatever I try, unless you run not CMake prior to compiling, it will try to build a Win32 DLL and fails.
So since I cannot find a way to provide .sln and .vcxproj files that can be used out of the box, the single .sln file is just senseless alone.