Gui: change string from Document to Documentation
The add property dialog (introduced by the LinkMerge) previously
showed "Document" which is confusing. The edit box is there
to provide an explanation on the new property, that is,
"Documentation".

Also improve the tooltips. Also change "Append" to "Prefix"
in the checkbox as it's more clear that the `Group` is added
at the beginning to form the full name of the property.